requirements.txt
# UI
streamlit

# Data handling
pandas
numpy
scikit-learn
python-dotenv

# Embeddings & Retrieval
sentence-transformers
faiss-cpu

# Groq library (if you are actually using Groq)
groq
